requirements.txt
streamlit>=1.30
plotly
pandas
spacy
pdfplumber
python-docx
scikit-learn
numpy
requests